Добавляет модуль карты и вносит правки
