+++ index.html
@@ -230,6 +230,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
+  <script src="vendor/pristine/pristine.min.js"></script>
   <script src="vendor/leaflet/leaflet.js"></script>
   <script src="vendor/nouislider/nouislider.js"></script>
   <script type="module" src="/js/main.js"></script>
