+++ js/main.js
@@ -1,4 +1,3 @@
-import {getAdvertisements} from './data.js';
 import './popup.js';
-getAdvertisements();
 import './user-form.js';
+import './map.js';
