+++ js/user-form.js
@@ -22,30 +22,6 @@ const disableForm = () => {
 
 disableForm();
 
-// ------- Активация формы -------
-
-const activateForm = () => {
-
-  advertisementForm.classList.remove('ad-form--disabled');
-
-  for (let i = 0; i < formElements.length; i++) {
-    formElements[i].removeAttribute('disabled');
-  }
-};
-
-activateForm();
-
-const activateFilters = () => {
-
-  mapForm.classList.remove('map__filters--disabled');
-
-  for (let i = 0; i < mapFilters.length; i++) {
-    mapFilters[i].removeAttribute('disabled');
-  }
-};
-
-activateFilters();
-
 // ------- Валидация формы -------
 
 const pristine = new Pristine(advertisementForm,
@@ -125,10 +101,34 @@ function getPriceErrorMessage () {
 
 pristine.addValidator(price, validateMinPrice, getPriceErrorMessage);
 
+// ------- Слайдер -------
+
+const priceSlider = document.querySelector('.ad-form__slider');
+
+noUiSlider.create(priceSlider, {
+  range: {
+    min: 0,
+    max: 100000,
+  },
+  start: 0,
+  step: 1,
+  connect: 'lower',
+});
+
+priceSlider.noUiSlider.on('update', () => {
+  price.value = priceSlider.noUiSlider.get();
+});
+
 function onTypeChange (evt) {
   price.placeholder = minPrice[evt.target.value];
   type = evt.target.value;
   pristine.validate(price);
+  priceSlider.noUiSlider.updateOptions({
+    range: {
+      min: minPrice[type],
+      max: 100000,
+    }
+  });
 }
 
 advertisementForm.querySelectorAll('[name="type"]').forEach((item) => item.addEventListener('change', onTypeChange));
@@ -160,135 +160,7 @@ advertisementForm.addEventListener('submit', (evt) => {
   pristine.validate();
 });
 
-// ------- Карта -------
-
-const map = L.map('map-canvas')
-.on('load', () =>
-  advertisementForm.classList.remove('ad-form--disabled'); // активация формы
-
-  for (let i = 0; i < formElements.length; i++) {
-    formElements[i].removeAttribute('disabled');
-  }
-
-  mapForm.classList.remove('map__filters--disabled');
-
-  for (let i = 0; i < mapFilters.length; i++) {
-    mapFilters[i].removeAttribute('disabled');
-  })
-  .setView({
-  lat: 35.6895,
-  lng: 139.692,
-}, 10);
-
-L.tileLayer(
-'http://tile2.maps.2gis.com/tiles?x={x}&y={y}&z={z}',
-  {
-    attribution: '&copy; Предоставлено <a href="https://dev.2gis.ru/">2ГИС</a>', // это правильно???
-  },
-).addTo(map);
-
-const mainPinIcon = L.icon({
-  iconUrl: './img/main-pin.svg',
-  iconSize: [52, 52],
-  iconAnchor: [26, 52],
-});
-
-
-const mainPinMarker = L.marker(
-  {
-    lat: 35.6895,
-    lng: 139.692,
-  },
-  {
-    draggable: true,
-    icon: mainPinIcon,
-  },
-);
-
-mainPinMarker.addTo(map);
-
-const address = advertisementForm.querySelector('[name="address"]');
-
-mainPinMarker.on('moveend', (evt) => {
-  address.value = evt.target.getLatLng() // так можно сделать? - Содержимое поля адреса должно всегда соответствовать координатам метки.
-});
-
-const icon = L.icon({
-  iconUrl: './img/pin.svg',
-  iconSize: [40, 40],
-  iconAnchor: [20, 40],
-});
-
-const createAdvertisementPopup = (advertisement) => {
-  const cardTemplate = document.querySelector('#card').content.querySelector('.popup');
-  const popup = cardTemplate.cloneNode(true);
-    if (offer.title) {
-      popup.querySelector('.popup__title').textContent = offer.title;
-    }
-    if (offer.address) {
-      popup.querySelector('.popup__text--address').textContent = offer.address;
-    }
-    if (offer.price) {
-      popup.querySelector('.popup__text--price').textContent = `${offer.price} ₽/ночь`;
-    }
-    if (offer.type) {
-      popup.querySelector('.popup__type').textContent = offerTypes[offer.type];
-    }
-    if (offer.rooms && offer.guests) {
-      popup.querySelector('.popup__text--capacity').textContent = `${offer.rooms} комнаты для ${offer.guests} гостей`;
-    }
-    if (offer.checkin && offer.checkout) {
-      popup.querySelector('.popup__text--time').textContent = `Заезд после ${offer.checkin}, выезд до ${offer.checkout}`;
-    }
-    if (offer.features) {
-      popup.querySelector('.popup__features').textContent = offer.features;
-    }
-    if (offer.description) {
-      popup.querySelector('.popup__description').textContent = offer.description;
-    }
-    if (offer.photos) {
-      popup.querySelector('.popup__photos').setAttribute('src', offer.photos);
-    }
-    if (author.avatar) {
-      popup.querySelector('.popup__avatar').setAttribute('src', author.avatar);
-    }
-
-    return popup;
-  };
-
-advertisements.forEach((advertisement) => {
-  const {lat, lng} = advertisement;
-  const marker = L.marker(
-    {
-      lat,
-      lng,
-    },
-    {
-      icon,
-    },
-  );
-
-  marker
-    .addTo(map)
-    .bindPopup(createAdvertisementPopup(advertisement));
-});
-
-// ------- Слайдер -------
-
-const priceSlider = document.querySelector('ad-form__slider');
-
-valueElement.value = 0;
-
-noUiSlider.create(priceSlider, {
-  range: {
-    min: minPrice[type], // так можно для разных типов жилья или нет?
-    max: 100000,
-  },
-  start: 0,
-  step: 1,
-  connect: 'lower',
-});
-
-priceSlider.noUiSlider.on('update', () => {
-  price.value = priceSlider.noUiSlider.get();
-});
+export { advertisementForm };
+export { formElements };
+export { mapForm };
+export { mapFilters };
